Refactored: Datadog::Workers --> Datadog::Core::Workers
