+++ spec/datadog/core/workers/interval_loop_spec.rb
@@ -1,15 +1,15 @@
 # typed: false
 require 'spec_helper'
 
-require 'ddtrace/worker'
-require 'ddtrace/workers/interval_loop'
+require 'datadog/core/worker'
+require 'datadog/core/workers/interval_loop'
 
-RSpec.describe Datadog::Workers::IntervalLoop do
+RSpec.describe Datadog::Core::Workers::IntervalLoop do
   context 'when included into a worker' do
     subject(:worker) { worker_class.new(&task) }
 
     let(:worker_class) do
-      Class.new(Datadog::Worker) { include Datadog::Workers::IntervalLoop }
+      Class.new(Datadog::Core::Worker) { include Datadog::Core::Workers::IntervalLoop }
     end
 
     let(:task) { proc { |*args| worker_spy.perform(*args) } }
