+++ spec/datadog/core/workers/queue_spec.rb
@@ -1,15 +1,15 @@
 # typed: false
 require 'spec_helper'
 
-require 'ddtrace/worker'
-require 'ddtrace/workers/queue'
+require 'datadog/core/worker'
+require 'datadog/core/workers/queue'
 
-RSpec.describe Datadog::Workers::Queue do
+RSpec.describe Datadog::Core::Workers::Queue do
   context 'when included into a worker' do
     subject(:worker) { worker_class.new(&task) }
 
     let(:worker_class) do
-      Class.new(Datadog::Worker) { include Datadog::Workers::Queue }
+      Class.new(Datadog::Core::Worker) { include Datadog::Core::Workers::Queue }
     end
 
     let(:task) { proc { |*args| worker_spy.perform(*args) } }
