+++ spec/ddtrace/configuration_spec.rb
@@ -403,7 +403,7 @@ RSpec.describe Datadog::Configuration do
     describe '#runtime_metrics' do
       subject(:runtime_metrics) { test_class.send(:components).runtime_metrics }
 
-      it { is_expected.to be_a_kind_of(Datadog::Workers::RuntimeMetrics) }
+      it { is_expected.to be_a_kind_of(Datadog::Core::Workers::RuntimeMetrics) }
       it { expect(runtime_metrics.enabled?).to be false }
       it { expect(runtime_metrics.running?).to be false }
     end
