+++ spec/spec_helper.rb
@@ -147,7 +147,7 @@ RSpec.configure do |config|
         # teardown in those tests.
         # They currently flood the output, making our test
         # suite output unreadable.
-        if example.file_path.start_with?('./spec/ddtrace/workers/')
+        if example.file_path.start_with?('./spec/datadog/core/workers/', './spec/ddtrace/workers/')
           puts # Add newline so we get better output when the progress formatter is being used
           RSpec.warning("FIXME: #{example.file_path}:#{example.metadata[:line_number]} is leaking threads")
           next
